requirements.txt
yt-dlp
moviepy
openai
google-auth google-auth-oauthlib google-auth-httplib2